+++ iNOPC_Server/Web/RequestTypes/IdOnly.cs
@@ -2,6 +2,6 @@
 {
     class IdOnly
     {
-        public int Id { get; set; } = 0;
+        public long Id { get; set; } = 0;
     }
 }
